+++ bot/src/on-pull-request.test.ts
@@ -9,6 +9,7 @@ test("empty", function () {
   const data: PullRequestContextData = {
     prettifierConfig: "",
     prettierConfig: "",
+    prettierIgnore: "",
   }
   const actual = parsePullRequestContextData(data)
   assert.deepEqual(actual.prettierConfig, {})
@@ -19,6 +20,7 @@ test("with content", function () {
   const data: PullRequestContextData = {
     prettifierConfig: "excludeFiles: dist",
     prettierConfig: "semi: false",
+    prettierIgnore: "",
   }
   const actual = parsePullRequestContextData(data)
   assert.deepEqual(actual.prettierConfig, { semi: false })