+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Prettifier changelog
 
+- 2020-04-07: handle `.prettierignore` files
 - 2020-04-07: Prettier 2.0 (formats slightly differently)
 - 2020-03-07: new website https://prettifier.io
 - 2020-02-13: GitHubOps
-- 2020-02-08: Configuration option to check pull requests only
+- 2020-02-08: option to limit fixing to branches with pull requests
   [#100](https://github.com/kevgo/prettifier/pull/100)
 - 2020-01-13: mono-repository [#81](https://github.com/kevgo/prettifier/pull/81)
 - 2020-01-11: Prettify all files in a single commit